icu_locid

Enum SubtagOrderingResult

pub enum SubtagOrderingResult<I> {
    Subtags(I),
    Ordering(Ordering),
}

๐Deprecated since 1.5.0: if you need this, please file an issue

The result of a subtag iterator comparison operation.

See Locale::strict_cmp_iter.

Examples

Check whether a stream of subtags contains two expected locales back-to-back:

use icu::locid::{locale, SubtagOrderingResult};
use std::cmp::Ordering;

let subtags = b"en-US-it-IT".split(|b| *b == b'-');
let locales = [locale!("en-US"), locale!("it-IT")];
let mut result = SubtagOrderingResult::Subtags(subtags);
for loc in locales.iter() {
    match result {
        SubtagOrderingResult::Subtags(it) => {
            result = loc.strict_cmp_iter(it);
        }
        SubtagOrderingResult::Ordering(ord) => break,
    }
}

assert_eq!(Ordering::Equal, result.end());

Variants

Subtags(I)

๐Deprecated since 1.5.0: if you need this, please file an issue

Potentially remaining subtags after the comparison operation.

Ordering(Ordering)

๐Deprecated since 1.5.0: if you need this, please file an issue

Resolved ordering between the locale object and the subtags.

Implementations

impl<I> SubtagOrderingResult<I>
where I: Iterator,

pub fn end(self) -> Ordering

Invoke this function if there are no remaining locale objects to chain in order to get a fully resolved Ordering.