Module icu_locid::extensions::private

Private Use Extensions is a list of extensions intended for private use.

Those extensions are treated as a pass-through, and no Unicode related behavior depends on them.

The main struct for this extension is Private which is a list of Subtags.

Examples

use icu::locid::extensions_private_subtag as subtag;
use icu::locid::{locale, Locale};

let mut loc: Locale = "en-US-x-foo-faa".parse().expect("Parsing failed.");

assert!(loc.extensions.private.contains(&subtag!("foo")));
assert_eq!(loc.extensions.private.iter().next(), Some(&subtag!("foo")));

loc.extensions.private.clear();

assert!(loc.extensions.private.is_empty());
assert_eq!(loc, locale!("en-US"));

Structs

Private

A list of Private Use Extensions as defined in Unicode Locale Identifier specification.

Subtag

A single item used in a list of Private extensions.