Struct icu_locid::extensions::transform::Fields

pub struct Fields(_);

A list of Key-Value pairs representing functional information about content transformations.

Here are examples of fields used in Unicode:

s0, d0 - Transform source/destination

t0 - Machine Translation

h0 - Hybrid Locale Identifiers

You can find the full list in Unicode BCP 47 T Extension section of LDML.

Examples

use icu::locid::extensions::transform::{Fields, Key, Value};

let key: Key = "h0".parse()
    .expect("Failed to parse a Key.");
let value: Value = "hybrid".parse()
    .expect("Failed to parse a Value.");
let fields: Fields = vec![(key, value)].into_iter().collect();

assert_eq!(&fields.to_string(), "h0-hybrid");

Implementations

impl Fields

pub const fn new() -> Self

Returns a new empty list of key-value pairs. Same as default(), but is const.

Examples

use icu::locid::extensions::transform::Fields;

assert_eq!(Fields::new(), Fields::default());

pub fn is_empty(&self) -> bool

Returns true if there are no fields.

Examples

use icu::locid::Locale;
use icu::locid::extensions::transform::Fields;

let loc1 = Locale::from_bytes(b"und-t-h0-hybrid").unwrap();
let loc2 = Locale::from_bytes(b"und-u-ca-buddhist").unwrap();

assert!(!loc1.extensions.transform.fields.is_empty());
assert!(loc2.extensions.transform.fields.is_empty());

pub fn clear(&mut self)

Empties the Fields list.

Examples

use icu::locid::extensions::transform::{Fields, Key, Value};

let key: Key = "h0".parse()
    .expect("Failed to parse a Key.");
let value: Value = "hybrid".parse()
    .expect("Failed to parse a Value.");
let mut fields: Fields = vec![(key, value)].into_iter().collect();

assert_eq!(&fields.to_string(), "h0-hybrid");

fields.clear();

assert_eq!(&fields.to_string(), "");

pub fn contains_key<Q>(&self, key: &Q) -> bool where
Key: Borrow<Q>,
Q: Ord,

Returns true if the list contains a Value for the specified Key.

Examples

use icu::locid::extensions::transform::{Fields, Key, Value};

let key: Key = "h0".parse()
    .expect("Failed to parse a Key.");
let value: Value = "hybrid".parse()
    .expect("Failed to parse a Value.");
let mut fields: Fields = vec![(key, value)].into_iter().collect();

let key: Key = "h0".parse()
    .expect("Failed to parse a Key.");
assert!(&fields.contains_key(&key));

pub fn get<Q>(&self, key: &Q) -> Option<&Value> where
Key: Borrow<Q>,
Q: Ord,

Returns a reference to the Value corresponding to the Key.

Examples

use icu::locid::extensions::transform::{Fields, Key, Value};

let key: Key = "h0".parse()
    .expect("Failed to parse a Key.");
let value: Value = "hybrid".parse()
    .expect("Failed to parse a Value.");
let mut fields: Fields = vec![(key, value)].into_iter().collect();

let key: Key = "h0".parse()
    .expect("Failed to parse a Key.");
assert_eq!(
    fields.get(&key).map(|v| v.to_string()),
    Some("hybrid".to_string())
);

pub fn set(&mut self, key: Key, value: Value) -> Option<Value>

Sets the specified keyword, returning the old value if it already existed.

Examples

use std::str::FromStr;
use std::string::ToString;
use icu::locid::transform_ext_key;
use icu::locid::Locale;
use icu::locid::extensions::transform::Key;
use icu::locid::extensions::transform::Value;

const D0_KEY: Key = transform_ext_key!("d0");
let lower = Value::from_str("lower").expect("valid extension subtag");
let casefold = Value::from_str("casefold").expect("valid extension subtag");

let mut loc: Locale = "en-t-hi-d0-casefold".parse()
    .expect("valid BCP-47 identifier");
let old_value = loc.extensions.transform.fields.set(
    D0_KEY,
    lower
);

assert_eq!(old_value, Some(casefold));
assert_eq!(loc, "en-t-hi-d0-lower");

pub fn retain_by_key<F>(&mut self, predicate: F) where
F: FnMut(&Key) -> bool,

Retains a subset of fields as specified by the predicate function.

Examples

use icu::locid::Locale;
use std::str::FromStr;

let mut loc: Locale = "und-t-h0-hybrid-d0-hex-m0-xml".parse().unwrap();

loc.extensions.transform.fields.retain_by_key(|k| k == "h0");
assert_eq!(loc, "und-t-h0-hybrid");

loc.extensions.transform.fields.retain_by_key(|k| k == "d0");
assert_eq!(loc, "und");