icu_locid::extensions::unicode

Struct Value

pub struct Value(/* private fields */);

A value used in a list of Keywords.

The value has to be a sequence of one or more alphanumerical strings separated by -. Each part of the sequence has to be no shorter than three characters and no longer than 8.

Examples

use icu::locid::extensions::unicode::{value, Value};
use writeable::assert_writeable_eq;

assert_writeable_eq!(value!("gregory"), "gregory");
assert_writeable_eq!(
    "islamic-civil".parse::<Value>().unwrap(),
    "islamic-civil"
);

// The value "true" has the special, empty string representation
assert_eq!(value!("true").to_string(), "");

Implementations

impl Value

pub fn try_from_bytes(input: &[u8]) -> Result<Self, ParserError>

A constructor which takes a utf8 slice, parses it and produces a well-formed Value.

Examples

use icu::locid::extensions::unicode::Value;

Value::try_from_bytes(b"buddhist").expect("Parsing failed.");

pub const fn try_from_single_subtag(subtag: &[u8]) -> Result<Self, ParserError>

Const constructor for when the value contains only a single subtag.

Examples

use icu::locid::extensions::unicode::Value;

Value::try_from_single_subtag(b"buddhist").expect("valid subtag");
Value::try_from_single_subtag(b"#####").expect_err("invalid subtag");
Value::try_from_single_subtag(b"foo-bar").expect_err("not a single subtag");