Struct icu_locid::LanguageIdentifier

pub struct LanguageIdentifier {
    pub language: Language,
    pub script: Option<Script>,
    pub region: Option<Region>,
    pub variants: Variants,
}

A core struct representing a Unicode BCP47 Language Identifier.

Examples

use icu::locid::LanguageIdentifier;

let li: LanguageIdentifier = "en-US".parse()
    .expect("Failed to parse.");

assert_eq!(li.language, "en");
assert_eq!(li.script, None);
assert_eq!(li.region.unwrap(), "US");
assert_eq!(li.variants.len(), 0);
assert_eq!(li, "en-US");

Parsing

Unicode recognizes three levels of standard conformance for any language identifier:

well-formed - syntactically correct

valid - well-formed and only uses registered language, region, script and variant subtags…

canonical - valid and no deprecated codes or structure.

At the moment parsing normalizes a well-formed language identifier converting _ separators to - and adjusting casing to conform to the Unicode standard.

Any bogus subtags will cause the parsing to fail with an error. No subtag validation is performed.

Examples

use icu::locid::LanguageIdentifier;

let li: LanguageIdentifier = "eN_latn_Us-Valencia".parse()
    .expect("Failed to parse.");

assert_eq!(li.language, "en");
assert_eq!(li.script.unwrap(), "Latn");
assert_eq!(li.region.unwrap(), "US");
assert_eq!(li.variants.get(0).unwrap(), "valencia");

Fields
language: Language

Language subtag of the language identifier.

script: Option<Script>

Script subtag of the language identifier.

region: Option<Region>

Region subtag of the language identifier.

variants: Variants

Variant subtags of the language identifier.

Implementations

impl LanguageIdentifier

pub fn from_bytes(v: &[u8]) -> Result<Self, ParserError>

A constructor which takes a utf8 slice, parses it and produces a well-formed LanguageIdentifier.

Examples

use icu::locid::LanguageIdentifier;

let li = LanguageIdentifier::from_bytes(b"en-US")
    .expect("Parsing failed.");

assert_eq!(li.to_string(), "en-US");

pub fn from_locale_bytes(v: &[u8]) -> Result<Self, ParserError>

A constructor which takes a utf8 slice which may contain extension keys, parses it and produces a well-formed LanguageIdentifier.

Examples

use icu::locid::LanguageIdentifier;

let li = LanguageIdentifier::from_locale_bytes(b"en-US-x-posix")
    .expect("Parsing failed.");

assert_eq!(li.to_string(), "en-US");

This method should be used for input that may be a locale identifier. All extensions will be lost.

pub const UND: Self = Self { language: subtags::Language::UND, script: None, region: None, variants: subtags::Variants::new(), }

The default undefined language “und”. Same as default().

Examples

use icu::locid::LanguageIdentifier;

assert_eq!(LanguageIdentifier::default(), LanguageIdentifier::UND);
assert_eq!("und", LanguageIdentifier::UND.to_string());

pub fn canonicalize<S: AsRef<[u8]>>(input: S) -> Result<String, ParserError>

This is a best-effort operation that performs all available levels of canonicalization.

At the moment the operation will normalize casing and the separator, but in the future it may also validate and update from deprecated subtags to canonical ones.

Examples

use icu::locid::LanguageIdentifier;

assert_eq!(LanguageIdentifier::canonicalize("pL_latn_pl"), Ok("pl-Latn-PL".to_string()));

pub fn cmp_bytes(&self, other: &[u8]) -> Ordering

Compare this LanguageIdentifier with a BCP-47 string.

The return value is equivalent to what would happen if you first converted this LanguageIdentifier to a BCP-47 string and then performed a byte comparison.

This function is case-sensitive and results in a total order, so it is appropriate for binary search. The only argument producing Ordering::Equal is self.to_string().

Examples

use icu::locid::LanguageIdentifier;
use std::cmp::Ordering;

let bcp47_strings: &[&[u8]] = &[
    b"pl-Latn-PL",
    b"und",
    b"und-Adlm",
    b"und-GB",
    b"und-ZA",
    b"und-fonipa",
    b"zh",
];

for ab in bcp47_strings.windows(2) {
    let a = ab[0];
    let b = ab[1];
    assert!(a.cmp(b) == Ordering::Less);
    let a_langid = LanguageIdentifier::from_bytes(a).unwrap();
    assert!(a_langid.cmp_bytes(b) == Ordering::Less);
}