Module icu_locid::extensions::other

Other Use Extensions is a list of extensions other than unicode, transform or private.

Those extensions are treated as a pass-through, and no Unicode related behavior depends on them.

The main struct for this extension is Other which is a list of Keys.

Examples

use icu::locid::Locale;
use icu::locid::extensions::other::Other;

let mut loc: Locale = "en-US-a-foo-faa".parse()
    .expect("Parsing failed.");

Structs

Key

A single item used in a list of Other extensions.

Other

A list of Other Use Extensions as defined in Unicode Locale Identifier specification.